feat: feat:added and displayed pokemon overview
- added dartX in pubspec and used it for capitalizing the first letter of pokemon name
- moved gridview back to overview page and created a widget for card instead
- fixed string manipulation for pokemon image url
- fixed constants
