+++ pokedex_async_redux/lib/features/pokemon_overview_page.dart
@@ -1,5 +1,6 @@
 import 'package:pokedex_async_redux/api/model/models.dart';
-import 'package:pokedex_async_redux/features/widget/pokemon_overview_page_gridview.dart';
+import 'package:pokedex_async_redux/features/widget/pokemon_overview_page_card.dart';
+import 'package:pokedex_async_redux/utilities/colors.dart';
 import 'package:pokedex_async_redux/utilities/constants.dart';
 import 'package:flutter/material.dart';
 
@@ -10,17 +11,25 @@ class PokemonOverviewPage extends StatelessWidget {
   }) : super(key: key);
 
   final List<Pokemon> pokemons;
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
       home: Scaffold(
         appBar: AppBar(
-          title: appBarTitle,
+          title: const Text(appBarTitle),
           backgroundColor: appDefaultColor,
         ),
         body: Center(
-          child: PokemonOverviewGridview(pokemons: pokemons),
+          child: GridView.builder(
+            gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 2),
+            itemCount: pokemons.length,
+            itemBuilder: (context, index) {
+              final pokemon = pokemons[index];
+              return PokemonCard(pokemon: pokemon);
+            },
+          ),
         ),
       ),
     );
