+++ pokedex_async_redux/lib/state/action/pokemon_actions.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:async_redux/async_redux.dart';
 import 'package:pokedex_async_redux/api/api_service.dart';
 import 'package:pokedex_async_redux/state/app_state.dart';
-import '../../utilities/string_constant.dart';
+import 'package:pokedex_async_redux/utilities/constants.dart';
 
 /// Setting of pokemons from pokemon api
 class GetPokemonsAction extends ReduxAction<AppState> {
