+++ pokedex_async_redux/lib/utilities/colors.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+const appDefaultColor = Color.fromARGB(255, 13, 39, 83);
