--- pokedex_async_redux/lib/utilities/int_constant.dart
@@ -1,2 +1,0 @@
-const numberOne = 1;
-const numberTwo = 2;
