--- pokedex_async_redux/lib/utilities/string_constant.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-const appBarTitle = Text("Pokedex");
-const pokemonActionOffset = "0";
-const pokemonActionLimit = "151";
-const homePageTitle = "Home Page";
-const pokemonImageUri = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/home/[index]";
-const pokemonImageExtention = ".png";
-const pokemonNameInput = 'charizard';
