+++ pokedex_async_redux/pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  dartx: ^1.1.0
 
   # architecture
   async_redux: ^18.0.0
